Fix vector operator in commit pols

+++ src/starkpil/src/polinomial.hpp
@@ -74,7 +74,7 @@ public:
     uint64_t length(void) { return _degree * _dim; }
     uint64_t size(void) { return _degree * _dim * sizeof(Goldilocks::Element); }
 
-    Goldilocks::Element *operator[](int i) { return &_pAddress[i * _offset]; };
+    Goldilocks::Element *operator[](uint64_t i) { return &_pAddress[i * _offset]; };
 
     std::string toString(uint numElements = 0, uint radix = 10)
     {
